Move methods to utils and add more tests
